fix: resolve key bugs, add editor prop for monsterType
