+++ src/components/GameEditor/NuzlockeGameTags.tsx
@@ -10,6 +10,7 @@ export interface NuzlockeGameTagsProps {
     game: Game;
     data: State;
     color: string;
+    isCopy: boolean;
 }
 
 export function NuzlockeGameTags({
@@ -18,29 +19,42 @@ export function NuzlockeGameTags({
     game,
     data,
     color,
+    isCopy,
 }: NuzlockeGameTagsProps) {
     return <>
-        <div style={{display: 'flex', flexDirection: 'column', width: '6rem', marginRight: '2rem', justifyContent: 'space-between', alignItems: 'space-between'}}>
-            <Tag round style={{
-                background: gameOfOriginToColor(game),
-                color: darkMode ? color : game === 'None' ? '#000' : color,
-            }}>{game}</Tag>
-            {isCurrent && <Tag round style={{
-                background: 'rgba(0,0,0,0.1)',
-                color: darkMode ? '#fff' : '#000',
-                marginTop: '2px',
-            }}>
-                Current
-            </Tag>}
-        </div>
         <div style={{
             display: 'flex',
             width: '20rem',
             justifyContent: 'center',
             alignItems: 'center',
             pointerEvents: 'none',
+            flexDirection: 'column',
+            margin: 'auto',
         }}>
-            {data?.pokemon?.filter(p => p.status === 'Team').map(poke => <PokemonIcon key={poke.id} {...poke} />)}
+            <div style={{
+                display: 'flex',
+            }}>{data?.pokemon?.filter(p => p.status === 'Team').map(poke => <PokemonIcon key={poke.id} {...poke} />)}</div>
+            <div style={{display: 'flex', justifyContent: 'space-between', alignItems: 'space-between', minWidth: '50%'}}>
+                <Tag round style={{
+                    margin: '0 2px',
+                    background: gameOfOriginToColor(game),
+                    color: darkMode ? color : game === 'None' ? '#000' : color,
+                }}>{game}</Tag>
+                {isCurrent && <Tag round style={{
+                    margin: '0 2px',
+                    background: 'rgba(0,0,0,0.1)',
+                    color: darkMode ? '#fff' : '#000',
+                }}>
+                    Current
+                </Tag>}
+                {isCopy && <Tag round style={{
+                    margin: '0 2px',
+                    background: 'rgba(0,0,0,0.1)',
+                    color: darkMode ? '#fff' : '#000',
+                }}>
+                    Copy
+                </Tag>}
+            </div>
         </div>
     </>;
 };
