+++ src/components/GameEditor/NuzlockeSave.tsx
@@ -1,0 +1,187 @@
+/* eslint-disable no-underscore-dangle */
+import * as React from 'react';
+import {
+    Popover,
+    Button,
+    Menu,
+    Position,
+    MenuItem,
+    Intent,
+    Icon,
+    Toaster,
+} from '@blueprintjs/core';
+import { connect } from 'react-redux';
+import { State } from 'state';
+import { updateNuzlocke, deleteNuzlocke, newNuzlocke, switchNuzlocke, replaceState, updateSwitchNuzlocke } from 'actions';
+import { gameOfOriginToColor, getContrastColor } from 'utils';
+import { omit } from 'ramda';
+import { createStore } from 'redux';
+import { appReducers } from 'reducers';
+import { NuzlockeGameTags } from './NuzlockeGameTags';
+
+export interface NuzlockeSaveControlsProps {
+    nuzlockes: State['nuzlockes'];
+    darkMode?: boolean;
+    state: string;
+    updateNuzlocke: updateNuzlocke;
+    deleteNuzlocke: deleteNuzlocke;
+    newNuzlocke: newNuzlocke;
+    switchNuzlocke: switchNuzlocke;
+    replaceState: replaceState;
+    updateSwitchNuzlocke: updateSwitchNuzlocke;
+}
+
+const sort = (a, b) => a.id - b.id;
+
+export class NuzlockeSaveBase extends React.Component<NuzlockeSaveControlsProps> {
+    // eslint-disable-next-line camelcase
+    public UNSAFE_componentWillMount() {
+        const {nuzlockes, newNuzlocke, state} = this.props;
+        if (!nuzlockes.currentId || nuzlockes.currentId === '') {
+            newNuzlocke(state, {isCopy: false});
+        }
+    }
+
+    public renderMenu() {
+
+        const {state, replaceState, updateSwitchNuzlocke, newNuzlocke, updateNuzlocke, deleteNuzlocke, switchNuzlocke, darkMode} = this.props;
+        const {nuzlockes} = this.props;
+        const {currentId} = this.props.nuzlockes;
+        const saves = nuzlockes.saves.sort(sort);
+
+        return <div style={{
+            padding: '0.5rem',
+        }}>
+            {saves.map(nuzlocke => {
+                const id = nuzlocke.id;
+                const {isCopy} = nuzlocke;
+                const isCurrent = currentId === id;
+                const data = nuzlocke.data;
+
+                if (!data || data === '{}' || data === '{ }') {
+                    return null;
+                }
+
+                let parsedData;
+
+                try {
+                    parsedData = isCurrent ? JSON.parse(state) : JSON.parse(data);
+                    //parsedData = JSON.parse(data);
+                } catch (e) {
+
+                }
+
+                if (!parsedData) {
+                    return null;
+                }
+
+                const game = parsedData?.game?.name;
+                const color = getContrastColor(gameOfOriginToColor(game));
+
+                return <div style={{
+                    display: 'flex',
+                    alignItems: 'center',
+                    border: darkMode ? '1px solid #444' : '1px solid #ccc',
+                    padding: '0.5rem',
+                    borderRadius: '0.25rem',
+                    boxShadow: '0 0 4px rgba(0,0,0,0.1)',
+                    marginBottom: '2px',
+                    justifyContent: 'space-between',
+                    // width: isMobile() ? '80%' : '100%',
+                }} key={id}>
+                    <NuzlockeGameTags
+                        darkMode={darkMode}
+                        game={game}
+                        color={color}
+                        data={parsedData}
+                        isCurrent={isCurrent}
+                        isCopy={isCopy}
+                    />
+                    <Popover position={Position.BOTTOM_RIGHT} content={<Menu><MenuItem shouldDismissPopover={false} disabled={isCurrent} icon='swap-horizontal' onClick={() => {
+                        try {
+                            console.log(
+                                currentId,
+                                id,
+                                parsedData.pokemon.map(p => p.species),
+                                JSON.parse(state).pokemon.map(p => p.species),
+                            );
+                            updateSwitchNuzlocke(currentId, id, state);
+                            // switchNuzlocke(id);
+                            replaceState(parsedData);
+                        } catch (e) {
+                            const toaster = Toaster.create();
+                            toaster.show({
+                                message: `Failed to switch nuzlockes. ${e}`,
+                                intent: Intent.DANGER,
+                            });
+                        }
+                    }} text='Switch to this Nuzlocke' />
+                    <MenuItem shouldDismissPopover={false} icon='clipboard'
+                        onClick={_ => {
+                            try {
+                                if (typeof data !== 'string') {
+                                    throw new Error('Data is not in correct format.');
+                                }
+                                newNuzlocke(data, {isCopy: true});
+                            } catch (e) {
+                                const toaster = Toaster.create();
+                                toaster.show({
+                                    message: `Failed to copy nuzlocke. ${e}`,
+                                    intent: Intent.DANGER,
+                                });
+                            }
+                        }}
+                        text='Copy this Nuzlocke'
+                    >
+                    </MenuItem>
+                    <MenuItem disabled={saves.length === 1} shouldDismissPopover={false} icon='trash' intent={Intent.DANGER} onClick={() => {
+                        try {
+                            deleteNuzlocke(id);
+                            if (isCurrent) {
+                                switchNuzlocke(saves[0].id);
+                                replaceState(JSON.parse(saves[0].data));
+                            }
+                        } catch (e) {
+                            const toaster = Toaster.create();
+                            toaster.show({
+                                message: `Failed to delete nuzlocke. ${e}`,
+                                intent: Intent.DANGER,
+                            });
+                        }
+                    }} text='Delete Nuzlocke' /></Menu>}>
+                        <Icon style={{ transform: 'rotate(90deg)', marginLeft: 'auto' }} icon="more" />
+                    </Popover>
+                </div>;
+            })}
+            <Button intent={Intent.SUCCESS} icon='add' onClick={() => {
+                updateNuzlocke(currentId, state);
+                const data = createStore(appReducers)?.getState();
+                newNuzlocke(JSON.stringify(data), {isCopy: false});
+                replaceState(data);
+            }}>
+                New Nuzlocke
+            </Button>
+
+        </div>;
+    }
+
+    public render() {
+        return this.renderMenu();
+    }
+}
+
+export const NuzlockeSave = connect(
+    (state: State) => ({
+        nuzlockes: state.nuzlockes,
+        state: JSON.stringify(omit(['nuzlockes'], state)),
+        darkMode: state.style.editorDarkMode,
+    }),
+    {
+        updateNuzlocke,
+        newNuzlocke,
+        deleteNuzlocke,
+        switchNuzlocke,
+        replaceState,
+        updateSwitchNuzlocke,
+    }
+)(NuzlockeSaveBase as any);
