+++ src/components/PokemonEditor/PokemonEditor.tsx
@@ -1,4 +1,4 @@
-import { Button, Icon, Spinner } from '@blueprintjs/core';
+import { Button, Callout, Icon, Intent, Spinner } from '@blueprintjs/core';
 import * as React from 'react';
 import { connect } from 'react-redux';
 
@@ -59,6 +59,7 @@ const PokemonLocationChecklist = ({
 
     const getLocIcon = (name) => {
         const poke = pokemon.find((p) => p.met === name && p.gameOfOrigin === game.name);
+        console.log(name, poke?.met, poke?.gameOfOrigin, game.name);
         if (poke && !poke.hidden) {
             return (
                 <>
@@ -95,10 +96,10 @@ const PokemonLocationChecklist = ({
                     </div>
                 );
             })}
-            <div>
+            <Callout intent={Intent.WARNING} style={{fontSize: '80%', marginTop: '0.5rem'}}>
                 Tip: Pokémon with the "hidden" attribute are a great option for including Pokemon
                 that got away on a certain route!
-            </div>
+            </Callout>
         </div>
     );
 };
