+++ src/models/Editor.ts
@@ -1,5 +1,6 @@
 export interface Editor {
     minimized: boolean;
     temtemMode?: boolean;
+    monsterType?: string;
     showResultInMobile?: boolean;
 }
