+++ src/reducers/editor.ts
@@ -2,7 +2,7 @@ import { CHANGE_EDITOR_SIZE, Action, TOGGLE_TEMTEM_MODE, TOGGLE_MOBILE_RESULT_VI
 import { Editor } from 'models';
 
 export function editor(
-    state: Editor = { minimized: false, temtemMode: false, showResultInMobile: false },
+    state: Editor = { minimized: false, temtemMode: false, showResultInMobile: false, monsterType: 'Pokémon' },
     action: Action<CHANGE_EDITOR_SIZE | TOGGLE_TEMTEM_MODE | TOGGLE_MOBILE_RESULT_VIEW>,
 ) {
     switch (action.type) {
@@ -14,6 +14,7 @@ export function editor(
             return {
                 ...state,
                 temtemMode: !state.temtemMode,
+                monsterType: !state.temtemMode ? 'TemTem' : 'Pokémon',
             };
         case TOGGLE_MOBILE_RESULT_VIEW:
             return {
