+++ src/utils/feature.ts
@@ -3,4 +3,7 @@ export const feature = Object.freeze({
     temTemMode: process.env.TEM_TEM_SUPPORT,
     multipleNuzlockes: process.env.MULTIPLE_NUZLOCKES,
     fileUploads: process.env.FILE_UPLOADS,
+    newHotkeys: process.env.NEW_HOTKEYS,
+    resultv2: process.env.RESULT_V2,
+    hallOfFame: process.env.HALL_OF_FAME,
 });
